backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.7.4
google-generativeai==0.8.0
celery==5.4.0
redis==5.0.8
httpx==0.27.2
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
